% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconstruct.filename.R
\name{deconstruct.filename}
\alias{deconstruct.filename}
\title{Deconstructs a given filename}
\usage{
deconstruct.filename(filename)
}
\arguments{
\item{image.filename}{Character vector. Filename to be deconstructed.}
}
\value{
Dataframe. Column 1 is ID, Column 2 is Datetime from filename.
}
\description{
Deconstructs a given filename
}
\details{
Extracts a (parochially defined) set of data from a filename. Not generalisable
}
\author{
Jamie Rylance, \email{jamie.rylance@lstmed.ac.uk}
}
\keyword{internal}
