% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageout.R
\name{imageout}
\alias{imageout}
\title{Produce graphic output of EasyLog data}
\usage{
imageout(df, image.filename)
}
\arguments{
\item{df}{Dataframe. Columns labelled 'x' which contains POSIT time data and 'y' which contains carbon monoxide data (in ppm)}

\item{image.filename}{Character vector. Filename for the output.}
}
\value{
Logical. TRUE if image successfully saved, FALSE if not.
}
\description{
Produce graphic output of EasyLog data
}
\details{
Produces a scatter plot of EasyLog data, and saves as a PDF.
}
\examples{

}
\author{
Jamie Rylance, \email{jamie.rylance@lstmed.ac.uk}
}
\keyword{internal}
