% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.easyLogBatch.R
\name{read.easyLogBatch}
\alias{read.easyLogBatch}
\title{Import and process data from multiple EasyLog files}
\usage{
read.easyLogBatch(filedir, outputdir = "output", imagedir = "output/images",
  noimage = TRUE)
}
\arguments{
\item{filedir}{Character vector. Directory containing EasyLog CSV files to process.}

\item{outputdir}{Optional. Character vector. Directory into which data will be output (default=current working directory).}

\item{imagedir}{Optional. Character vector. Only used if noimage=FALSE. Directory into which image files will be output (default=current working directory).}

\item{noimage}{Logical. If TRUE, no images are produced (default=TRUE).}
}
\value{
Dataframe of test results
}
\description{
Import and process data from multiple EasyLog files
}
\details{
To extract data from multiple EasyLog files and perform summary calculations
}
\examples{
read.easyLogBatch("example/data", "example/output", "example/output/images", noimage=FALSE)

read.easyLogBatch("example/data", "example/output", noimage=TRUE)

}
\author{
Jamie Rylance, \email{jamie.rylance@lstmed.ac.uk}
}
