% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.easyLog.R
\name{read.easyLog}
\alias{read.easyLog}
\title{Import and process data from a single EasyLog file}
\usage{
read.easyLog(filename, imagedir = ".", noimage = TRUE)
}
\arguments{
\item{filename}{Character vector. Filename of the EasyLog CSV file to process.}

\item{imagedir}{Optional. Character vector. Only used if noimage=FALSE. Directory into which image files will be output (default=current working directory)}

\item{noimage}{Logical. If TRUE, no images are produced (default=TRUE).}
}
\value{
Dataframe of test results
}
\description{
Import and process data from a single EasyLog file
}
\details{
To extract data from an EasyLog file and perform summary calculations
}
\examples{
read.easyLog("example/data/example.txt", noimage=TRUE)

read.easyLog("/example/data/example.txt", imagedir="/example/images", noimage=FALSE)

}
\author{
Jamie Rylance, \email{jamie.rylance@lstmed.ac.uk}
}
